refactor(privilege): move thorUserId out of the shizuku package
Review point: `thorUserId` landed in `data/source/local/shizuku/` because
that is where the first call site was, but neither privilege mode owns
the question "which user is Thor". Leaving it there made the dhizuku
package import from the shizuku package for a value both need equally.

It now lives in `data/source/local/ThorUser.kt`, the package both helpers
sit under. That keeps the dependency direction the codebase already has —
`data/gateway` imports from `data/source/local` and nothing imports back
— which putting it beside `userIdOf` in `data/gateway/AndroidUserIds.kt`
would have inverted. The two files cross-reference each other: `userIdOf`
maps some other uid to its user, `thorUserId` is Thor's own.

While moving it, folded in two copies of its definition that were already
inlined in `Shizuku.kt` — `clearCache` and `clearAppData` each spelled out
`android.os.Process.myUserHandle().hashCode()`. Same expression, so no
behaviour change; it is the same one-symbol argument this PR opens with.

497 tests, 0 failures; lint unchanged (5 pre-existing VectorPath hints).
