+++ app/src/main/java/com/valhalla/thor/data/source/local/shizuku/Packages.kt
@@ -8,24 +8,9 @@ import android.app.AppOpsManager
 import android.content.Context
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
-import android.os.Process
 import androidx.core.content.getSystemService
 import com.valhalla.bypass.Bypass
-
-/**
- * The Android user Thor itself runs as — the one every per-package command has to name.
- *
- * Read in-process from [Process.myUserHandle]: no shell, no permission, no binder. Both privilege
- * helpers resolve the user through this one symbol so that a freeze rung and its fallback cannot
- * name different users; a rung that disables for user 10 followed by a fallback that uninstalls for
- * user 0 leaves the verify confirming a state nobody set. `SUSPEND_USER_ID` in `RootSystemGateway`
- * makes the same point from the other direction.
- *
- * Deliberately **not** `am get-current-user`, which is what the two helpers used to shell out for.
- * That reports the *foreground* user — a different number from this one on any work-profile device
- * — and it is a command Dhizuku's device-owner identity is not permitted to run at all.
- */
-internal val thorUserId: Int get() = Process.myUserHandle().hashCode()
+import com.valhalla.thor.data.source.local.thorUserId
 
 class Packages(private val app: Context) {
